add views + create post
